Install NPM3 in baseimage
NPM 2, the default installed with Node.js 4, is very old and has quite a
few bugs. One of them is that it does not properly forward signals, such
as `SIGTERM` to its child processes.

When `docker stop` is run, docker sends the `SIGTERM` signal to the topmost
process in the Docker container. It then waits (a default of) [10
seconds](https://docs.docker.com/engine/reference/commandline/stop/)
before sending `SIGKILL`.

This means, without upgrading to NPM3, our docker containers will hang
for 10 seconds whenever we try to stop them.

Ideally, we will upgrade to an even newer version of NPM in the near
future. NPM 4 and especially 5 are quite a bit faster and fix even more
bugs.