Add lipbostal service
The Pelias API has now been modified to expect libpostal to live as a
separate HTTP service rather than being included in memory.

This PR adds the libpostal service to `docker-compose.yml` but doesn't
change anything else. The current (as of 2018/3/19) `production` version
of Pelias API does not yet use the libpostal service. It's convenient to
have this merged so that we can start the process of switching over,
though.

Connects https://github.com/pelias/dockerfiles/pull/42